[PS-25] refactor: 좌표 validation 한국 범위로 변경
