+++ src/main/java/com/tf4/photospot/global/argument/CoordinateValidator.java
@@ -1,17 +1,23 @@
-package com.tf4.photospot.global.constraint;
+package com.tf4.photospot.global.argument;
 
 import com.tf4.photospot.global.dto.CoordinateDto;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-public class CoordinateValidator implements ConstraintValidator<Coordinate, CoordinateDto> {
+/*
+ * WGS84 좌표
+ * 한국
+ * 위도 약 33 ~ 39
+ * 경도 약 124 ~ 132
+ * */
+public class CoordinateValidator implements ConstraintValidator<KoreaCoordinate, CoordinateDto> {
 	private static final String COORD_NOT_EMPTY = "좌표에 빈 값이 들어갈 수 없습니다.";
 	private static final String COORD_INVALID_RANGE = "좌표의 범위가 유효하지 않습니다.";
-	private static final double MIN_LATITUDE = -90.0;
-	private static final double MAX_LATITUDE = 90.0;
-	private static final double MIN_LONGITUDE = -180.0;
-	private static final double MAX_LONGITUDE = 180.0;
+	private static final double MIN_LATITUDE = 33.0;
+	private static final double MAX_LATITUDE = 39.0;
+	private static final double MIN_LONGITUDE = 124.0;
+	private static final double MAX_LONGITUDE = 132.0;
 
 	@Override
 	public boolean isValid(CoordinateDto value, ConstraintValidatorContext context) {
