+++ src/main/java/com/tf4/photospot/global/argument/KoreaCoordinate.java
@@ -1,4 +1,4 @@
-package com.tf4.photospot.global.constraint;
+package com.tf4.photospot.global.argument;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -11,7 +11,7 @@ import jakarta.validation.Payload;
 @Constraint(validatedBy = CoordinateValidator.class)
 @Target({ElementType.FIELD, ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Coordinate {
+public @interface KoreaCoordinate {
 	String message() default "잘못된 좌표입니다.";
 
 	Class<?>[] groups() default {};
