+++ src/main/java/com/tf4/photospot/global/dto/CoordinateDto.java
@@ -2,10 +2,10 @@ package com.tf4.photospot.global.dto;
 
 import org.locationtech.jts.geom.Point;
 
-import com.tf4.photospot.global.constraint.Coordinate;
+import com.tf4.photospot.global.argument.KoreaCoordinate;
 import com.tf4.photospot.global.util.PointConverter;
 
-@Coordinate
+@KoreaCoordinate
 public record CoordinateDto(
 	Double lon,
 	Double lat
